requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.2
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
Jinja2==3.1.6
libsql==0.1.11
libsql-experimental==0.0.55
MarkupSafe==3.0.3
pillow==11.3.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlalchemy-libsql==0.2.0
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
